+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # DEV
+- Add wagtailnest to apps if not detected by dj-core
 # 0.0.6
 - Add .pyup.yml
 - Update user forms for wagtail 1.12